+++ admin/src/Controllers/Pages.php
@@ -6,7 +6,7 @@ use Formwork\Admin\Exceptions\TranslatedException;
 use Formwork\Admin\Fields\Fields;
 use Formwork\Admin\Uploader;
 use Formwork\Admin\Utils\JSONResponse;
-use Formwork\Admin\Utils\LanguageCodes;
+use Formwork\Languages\LanguageCodes;
 use Formwork\Core\Formwork;
 use Formwork\Core\Page;
 use Formwork\Core\Site;
